внес правки по итогам ревью
Перенес функцию показа сообщения об ошибке получения данных с сервера из утилитарного модуля в модуль fetch, переписал функцию показа сообщений об отправке данных на сервер. А также поправил отрисовку объявлений на карте, теперь попапы работают исправно.
