+++ js/util.js
@@ -1,5 +1,3 @@
-const ALERT_SHOW_TIME = 10000;
-
 const getRandomInteger = (min, max) => {
   if (min >= 0 && max > min) {
     min = Math.ceil(min);
@@ -16,29 +14,8 @@ const getRandomFraction = (min, max, fractionalSymbolsQuantity) => {
   return 'Invalid range!';
 };
 
-const showAlert = (message) => {
-  const alertContainer = document.createElement('div');
-  alertContainer.style.zIndex = 100;
-  alertContainer.style.position = 'absolute';
-  alertContainer.style.left = 0;
-  alertContainer.style.top = 100;
-  alertContainer.style.right = 0;
-  alertContainer.style.padding = '10px 3px';
-  alertContainer.style.fontSize = '30px';
-  alertContainer.style.textAlign = 'center';
-  alertContainer.style.backgroundColor = 'red';
-
-  alertContainer.textContent = message;
-
-  document.body.append(alertContainer);
-
-  setTimeout(() => {
-    alertContainer.remove();
-  }, ALERT_SHOW_TIME);
-};
-
 const isEscEvent = (evt) => {
   return evt.key === ('Escape' || 'Esc');
 };
 
-export {getRandomInteger, getRandomFraction, showAlert, isEscEvent};
+export {getRandomInteger, getRandomFraction, isEscEvent};
